pyproject.toml
[project]
name = "mastodon-rss"
version = "0.1.0"
description = "My Favs and Bookmarks"
authors = [
    {name = "Stephan Heuel", email = "mail@ping13.net"}
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "click>=8.1.7",
    "feedgen>=1.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "envyaml>=1.10.211231",
    "transformers",
    "torch>=2.5.1",
    "html2text>=2024.2.26",
    "diskcache>=5.6.3",
    "bs4>=0.0.2",
    "pytest>=8.0.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }

